# Optimize CassandraService::getKnownPort