--- a/atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraServiceTest.java
+++ b/atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraServiceTest.java
@@ -33,6 +33,7 @@ import com.palantir.atlasdb.util.MetricsManagers;
 import com.palantir.refreshable.Refreshable;
 import java.net.InetSocketAddress;
 import java.net.UnknownHostException;
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;
@@ -267,6 +268,17 @@ public class CassandraServiceTest {
                         .satisfies(server -> assertThat(desired).contains(server));
             }
         }
+    }
+
+    @Test
+    public void ports() throws Exception {
+        assertThat(CassandraService.onlyPort(List.of(1, 1, 1).iterator())).isEqualTo(1);
+        assertThat(CassandraService.onlyPort(
+                        IntStream.generate(() -> 4224).boxed().limit(15).iterator()))
+                .isEqualTo(4224);
+        assertThatThrownBy(() -> CassandraService.onlyPort(List.of(1, 2, 1).iterator()))
+                .isInstanceOf(UnknownHostException.class)
+                .hasMessageContaining("No single known port");
     }
 
     private Set<CassandraServer> getRecommendedHostsFromAThousandTrials(