--- a/atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraServiceTest.java
+++ b/atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraServiceTest.java
@@ -272,11 +272,12 @@ public class CassandraServiceTest {
 
     @Test
     public void ports() throws Exception {
-        assertThat(CassandraService.onlyPort(List.of(1, 1, 1).iterator())).isEqualTo(1);
-        assertThat(CassandraService.onlyPort(
-                        IntStream.generate(() -> 4224).boxed().limit(15).iterator()))
-                .isEqualTo(4224);
-        assertThatThrownBy(() -> CassandraService.onlyPort(List.of(1, 2, 1).iterator()))
+        assertThat(CassandraService.onlyPort(List.of(1, 1, 1))).isEqualTo(1);
+        assertThat(IntStream.generate(() -> 4224).boxed().limit(15).collect(Collectors.toList()))
+                .hasSize(15)
+                .satisfies(cluster ->
+                        assertThat(CassandraService.onlyPort(cluster)).isEqualTo(4224));
+        assertThatThrownBy(() -> CassandraService.onlyPort(List.of(1, 2, 1)))
                 .isInstanceOf(UnknownHostException.class)
                 .hasMessageContaining("No single known port");
     }